+++ InvenTree/company/forms.py
@@ -8,6 +8,8 @@ from __future__ import unicode_literals
 from InvenTree.forms import HelperForm
 
 from .models import Company
+from .models import SupplierPart
+from .models import SupplierPriceBreak
 
 
 class EditCompanyForm(HelperForm):
@@ -36,4 +38,38 @@ class CompanyImageForm(HelperForm):
         model = Company
         fields = [
             'image'
+        ]
+
+
+class EditSupplierPartForm(HelperForm):
+    """ Form for editing a SupplierPart object """
+
+    class Meta:
+        model = SupplierPart
+        fields = [
+            'part',
+            'supplier',
+            'SKU',
+            'description',
+            'manufacturer',
+            'MPN',
+            'URL',
+            'note',
+            'base_cost',
+            'multiple',
+            'minimum',
+            'packaging',
+            'lead_time'
+        ]
+
+
+class EditPriceBreakForm(HelperForm):
+    """ Form for creating / editing a supplier price break """
+
+    class Meta:
+        model = SupplierPriceBreak
+        fields = [
+            'part',
+            'quantity',
+            'cost'
         ]
