+++ InvenTree/company/urls.py
@@ -36,3 +36,23 @@ company_urls = [
     # Redirect any other patterns
     url(r'^.*$', RedirectView.as_view(url='', permanent=False), name='company-index'),
 ]
+
+price_break_urls = [
+    url('^new/', views.PriceBreakCreate.as_view(), name='price-break-create'),
+
+    url(r'^(?P<pk>\d+)/edit/', views.PriceBreakEdit.as_view(), name='price-break-edit'),
+    url(r'^(?P<pk>\d+)/delete/', views.PriceBreakDelete.as_view(), name='price-break-delete'),
+]
+
+supplier_part_detail_urls = [
+    url(r'edit/?', views.SupplierPartEdit.as_view(), name='supplier-part-edit'),
+    url(r'delete/?', views.SupplierPartDelete.as_view(), name='supplier-part-delete'),
+
+    url('^.*$', views.SupplierPartDetail.as_view(), name='supplier-part-detail'),
+]
+
+supplier_part_urls = [
+    url(r'^new/?', views.SupplierPartCreate.as_view(), name='supplier-part-create'),
+
+    url(r'^(?P<pk>\d+)/', include(supplier_part_detail_urls)),
+]
