+++ InvenTree/part/__init__.py
@@ -5,6 +5,5 @@ It includes models for:
 
 - PartCategory
 - Part
-- SupplierPart
 - BomItem
 """
