+++ InvenTree/part/admin.py
@@ -3,8 +3,6 @@ from import_export.admin import ImportExportModelAdmin
 
 from .models import PartCategory, Part
 from .models import PartAttachment, PartStar
-from .models import SupplierPart
-from .models import SupplierPriceBreak
 from .models import BomItem
 
 
@@ -32,14 +30,6 @@ class BomItemAdmin(ImportExportModelAdmin):
     list_display = ('part', 'sub_part', 'quantity')
 
 
-class SupplierPartAdmin(ImportExportModelAdmin):
-    list_display = ('part', 'supplier', 'SKU')
-
-
-class SupplierPriceBreakAdmin(ImportExportModelAdmin):
-    list_display = ('part', 'quantity', 'cost')
-
-
 """
 class ParameterTemplateAdmin(admin.ModelAdmin):
     list_display = ('name', 'units', 'format')
@@ -54,5 +44,4 @@ admin.site.register(PartCategory, PartCategoryAdmin)
 admin.site.register(PartAttachment, PartAttachmentAdmin)
 admin.site.register(PartStar, PartStarAdmin)
 admin.site.register(BomItem, BomItemAdmin)
-admin.site.register(SupplierPart, SupplierPartAdmin)
-admin.site.register(SupplierPriceBreak, SupplierPriceBreakAdmin)
+
