+++ InvenTree/part/forms.py
@@ -11,8 +11,6 @@ from django import forms
 
 from .models import Part, PartCategory, PartAttachment
 from .models import BomItem
-from .models import SupplierPart
-from .models import SupplierPriceBreak
 
 
 class PartImageForm(HelperForm):
@@ -140,37 +138,3 @@ class EditBomItemForm(HelperForm):
 
         # Prevent editing of the part associated with this BomItem
         widgets = {'part': forms.HiddenInput()}
-
-
-class EditSupplierPartForm(HelperForm):
-    """ Form for editing a SupplierPart object """
-
-    class Meta:
-        model = SupplierPart
-        fields = [
-            'part',
-            'supplier',
-            'SKU',
-            'description',
-            'manufacturer',
-            'MPN',
-            'URL',
-            'note',
-            'base_cost',
-            'multiple',
-            'minimum',
-            'packaging',
-            'lead_time'
-        ]
-
-
-class EditPriceBreakForm(HelperForm):
-    """ Form for creating / editing a supplier price break """
-
-    class Meta:
-        model = SupplierPriceBreak
-        fields = [
-            'part',
-            'quantity',
-            'cost'
-        ]
